[US4][US5][T1][T2]
Controller et View Project terminé.

+++ ScrumDev/app/controllers/ProjectController.php
@@ -13,15 +13,47 @@ class ProjectController extends ControllerBase
     public function indexAction()
     {
 	 
-	$id_user=$this->session->get("auth");
-	$id_project=Member::findFirstByid_user($id_user);
-	$project=Project::findFirstByid_project($id_project);
+	 $numberPage = 1;
+        if ($this->request->isPost()) {
+            $query = Criteria::fromInput($this->di, "Project", $_POST);
+            $this->persistent->parameters = $query->getParams();
+        } else {
+            $numberPage = $this->request->getQuery("page", "int");
+        }
+
+        $parameters = $this->persistent->parameters;
+        if (!is_array($parameters)) {
+            $parameters = array();
+        }
+        $parameters["order"] = "id_project";
+
+        $project = Project::find($parameters);
+
+        $paginator = new Paginator(array(
+            "data" => $project,
+            "limit"=> 10,
+            "page" => $numberPage));
+
+			 $this->view->page = $paginator->getPaginate();
 
     }
+	
+	public function showAction($id_project)
+	{
+		$project=Project::findFirstByid_project($id_project);
+		$member = Member::query()
+		->where("id_project = :idpro:")
+		->andWhere("id_user=:iduser:")
+		->bind(array("idpro" => $id_project,"iduser"=>$this->session->get("auth")))
+		->execute();
+		$this->view->member=$member;
+		$this->view->project=$project;
+			
+	}
 
     /**
      * Searches for project
-     /
+     */
     public function searchAction()
     {
   
@@ -71,10 +103,18 @@ class ProjectController extends ControllerBase
      * Edits a project
      *
      * @param string $id_project
-     /
+     */
     public function editAction($id_project)
     {
-
+		$member = Member::findFirst(array('id_user = ?0 and id_project = ?1', 'bind' => array($this->session->get("auth"), $id_project)));
+		
+        if (!$member) {
+            return $this->dispatcher->forward(array(
+                "controller" => "project",
+                "action" => "index"
+            ));
+        }
+		
         if (!$this->request->isPost()) {
 
             $project = Project::findFirstByid_project($id_project);
@@ -132,9 +172,10 @@ class ProjectController extends ControllerBase
 		
 		$member->id_project=$project->id_project;
 		$member->id_user=$this->session->get('auth');
+		$member->type=0; 
 		
 		$member->save();
-
+		
        $this->flash->success("Le Projet a été créé");
 
         return $this->dispatcher->forward(array(
@@ -147,7 +188,7 @@ class ProjectController extends ControllerBase
     /**
      * Saves a project edited
      *
-     /
+     */
     public function saveAction()
     {
 
@@ -201,11 +242,12 @@ class ProjectController extends ControllerBase
      * Deletes a project
      *
      * @param string $id_project
-     /
+     */
     public function deleteAction($id_project)
     {
 
         $project = Project::findFirstByid_project($id_project);
+		$member = Member::find(array(' id_project = ?0', 'bind' => array($id_project)));
         if (!$project) {
             $this->flash->error("project was not found");
 
@@ -215,6 +257,18 @@ class ProjectController extends ControllerBase
             ));
         }
 
+		if (!$member->delete()) {
+
+            foreach ($project->getMessages() as $message) {
+                $this->flash->error($message);
+            }
+
+            return $this->dispatcher->forward(array(
+                "controller" => "project",
+                "action" => "search"
+            ));
+        }
+		
         if (!$project->delete()) {
 
             foreach ($project->getMessages() as $message) {
@@ -226,14 +280,16 @@ class ProjectController extends ControllerBase
                 "action" => "search"
             ));
         }
-
+		
+		
+		
         $this->flash->success("project was deleted successfully");
 
         return $this->dispatcher->forward(array(
             "controller" => "project",
             "action" => "index"
         ));
-    }*/
+    }
 
 
 }
